Revert "Merge branch 'develop' into feature/Make-it-easier-for-participants-to-select-unlocked-avatar-parts"
This reverts commit 57b8a912a1558280c78c383ef9f40d00dbb3718b, reversing
changes made to ed4f867f5f232646825706f9144dce72c4313f6a.
