--- src/GRA.Domain.Model/JobDetailsAvatarImport.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace GRA.Domain.Model
-{
-    public class JobDetailsAvatarImport
-    {
-        public string AssetPath { get; set; }
-    }
-}
