+++ src/GRA.Domain.Model/JobType.cs
@@ -4,7 +4,6 @@
     {
         SendBulkEmails = 1,
         RunReport,
-        UpdateVendorStatus,
-        AvatarImport
+        UpdateVendorStatus
     }
 }
