+++ src/GRA.Web/Views/Profile/Books.cshtml
@@ -61,10 +61,9 @@
                             <input asp-for="PaginateModel.CurrentPage" type="hidden" />
                             <input asp-for="Book.Id" type="hidden" />
                             <input asp-for="Book.Title" type="hidden" />
-                            <button type="submit" class="btn btn-danger btn-spinner pull-right" aria-label="Confirm">
+                            <button type="submit" class="btn btn-danger pull-right" aria-label="Confirm">
                                 <span class="fa fa-remove" aria-hidden="true"></span>
                                 @SharedLocalizer[GRA.Annotations.Interface.Delete]
-                                <span class="fa fa-spinner fa-pulse fa-lg fa-fw hidden"></span>
                             </button>
                         </form>
                     </div>
@@ -88,7 +87,7 @@
                     <div class="col-xs-12">
                         @if (Model.Books.Count() == 0)
                         {
-                            <div class="alert alert-warning">@SharedLocalizer[GRA.Annotations.Info.NoBooksLogged]</div>
+                        <div class="alert alert-warning">@SharedLocalizer[GRA.Annotations.Info.NoBooksLogged]</div>
                         }
                         else
                         {
